[tech debt] Reduce complexity in CategoriesStore to improve SRP
With the current change we reduce complexity in CategoriesStore, by introducing AreaStore which will provide DB access to AreaRepository with SRP principle in mind.
